@@ -1,25 +1,22 @@
 name: apn_widgets
 description: A carefully curated collection of widget made by and used within Appnormal projects
-version: 1.1.10
+version: 2.0.0-beta
 homepage: https://appnormal.com
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: "^2.0.1"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  apn_http:
-  apn_analytics:
-
-  cupertino_icons: ^1.0.0
-  keyboard_visibility: ^0.5.6
-  app_settings: ^4.0.4
-  flare_flutter: ^2.0.6
-  image_picker: ^0.6.7+12
-  intl:
+  cupertino_icons: ^1.0.2
+  flutter_keyboard_visibility: ^5.0.0
+  app_settings: ^4.1.0
+  flare_flutter: ^3.0.0-nullsafety.1
+  image_picker: ^0.7.3
+  intl: ^0.17.0
 
 dev_dependencies:
   flutter_test:
