@@ -1,4 +1,3 @@
-import 'package:apn_http/apn_http.dart';
 import 'package:apn_widgets/apn_widgets.dart';
 import 'package:example/features/shared/example_page.dart';
 import 'package:flutter/material.dart';
@@ -22,13 +21,13 @@ class _DialogsPageBody extends StatelessWidget {
       child: Column(
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
-          FlatButton(
+          TextButton(
               child: Text('Show alert'),
               onPressed: () => DialogPresenter.of(context).showAlert('Alert title', 'Alert message')),
-          FlatButton(
+          TextButton(
               child: Text('Show error'),
               onPressed: () => DialogPresenter.of(context).showError(ErrorResponse.fromMessage('This is an error'))),
-          FlatButton(
+          TextButton(
               child: Text('Show open app settings'),
               onPressed: () => DialogPresenter.of(context).showOpenAppSettingsAlert(
                   'We want to open app settings', 'Pretty please let us open the app settings')),
@@ -36,4 +35,15 @@ class _DialogsPageBody extends StatelessWidget {
       ),
     );
   }
+}
+
+class ErrorResponse {
+  ErrorResponse(this.message);
+
+  final String message;
+
+  factory ErrorResponse.fromMessage(String message) => ErrorResponse(message);
+
+  @override
+  String toString() => message;
 }
