@@ -5,29 +5,29 @@ import 'extensions.dart';
 
 /// Used primarily for cards, ripple on Android and animated pressed color on iOS
 class TappableOverlay extends StatefulWidget {
-  final VoidCallback onTap;
+  final VoidCallback? onTap;
 
   // The color of the ripple on Android
-  final Color highlightColor;
+  final Color? highlightColor;
 
   // The pressed color on iOS also used on Android if highLight color is null.
   // If this value is also null color will be used as a base color
-  final Color pressedColor;
+  final Color? pressedColor;
   final Widget child;
-  final double width;
-  final double height;
-  final EdgeInsets margin;
-  final BorderRadius borderRadius;
-  final HitTestBehavior behaviour;
+  final double? width;
+  final double? height;
+  final EdgeInsets? margin;
+  final BorderRadius? borderRadius;
+  final HitTestBehavior? behaviour;
   final bool disableIosTappable;
-  final ShapeBorder shape;
+  final ShapeBorder? shape;
 
   final bool expandWidth, expandHeight;
 
   const TappableOverlay({
-    Key key,
-    @required this.child,
-    @required this.onTap,
+    Key? key,
+    required this.child,
+    required this.onTap,
     this.highlightColor,
     this.pressedColor,
     this.width,
@@ -52,7 +52,7 @@ class TappableOverlay extends StatefulWidget {
 
 class _TappableOverlayState extends State<TappableOverlay> {
   bool pressed = false;
-  Size childSize;
+  Size? childSize;
 
   @override
   Widget build(BuildContext context) {
@@ -60,7 +60,7 @@ class _TappableOverlayState extends State<TappableOverlay> {
     final pressedColor = (widget.pressedColor ?? Colors.white).withOpacity(0.3);
     final highlightColor = widget.highlightColor ?? pressedColor.darken().withOpacity(0.2);
 
-    var child = widget.child;
+    Widget child = widget.child;
 
     if (widget.expandHeight) {
       child = Column(children: [Expanded(child: child)]);
@@ -150,9 +150,9 @@ class _MeasureSize extends StatefulWidget {
   final OnWidgetSizeChange onChange;
 
   const _MeasureSize({
-    Key key,
-    @required this.onChange,
-    @required this.child,
+    Key? key,
+    required this.onChange,
+    required this.child,
   }) : super(key: key);
 
   @override
@@ -186,7 +186,7 @@ class _MeasureSizeState extends State<_MeasureSize>  {
   }
 
   void measureSizeNextFrame() {
-    WidgetsBinding.instance.addPostFrameCallback(measureWidget);
+    WidgetsBinding.instance?.addPostFrameCallback(measureWidget);
   }
 
   void measureWidget(_) {
@@ -201,4 +201,4 @@ class _MeasureSizeState extends State<_MeasureSize>  {
   }
 }
 
-typedef void OnWidgetSizeChange(Size size);
+typedef void OnWidgetSizeChange(Size? size);
